pyproject.toml
[project]
name = "automaton-auditor"
version = "0.1.0"
description = "Week 2 Automaton Auditor: forensic detective swarm (repo + doc) with LangGraph, uv-managed deps and .env for API keys."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.9",
    "langsmith>=0.7.6",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pypdf>=5.0.0",
]
